requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
backcall=0.1.0=py37_0
blas=2.14=openblas
bzip2=1.0.8=h516909a_2
ca-certificates=2019.11.27=0
cairo=1.16.0=hfb77d84_1002
certifi=2019.11.28=py37_0
cffi=1.13.2=py37h2e261b9_0
cudatoolkit=10.0.130=0
cycler=0.10.0=pypi_0
dbus=1.13.6=he372182_0
decorator=4.4.1=py_0
expat=2.2.5=he1b5a44_1004
ffmpeg=4.1.3=h167e202_0
fontconfig=2.13.1=h86ecdb6_1001
freetype=2.9.1=h8a8886c_1
gettext=0.19.8.1=hc5be6a0_1002
giflib=5.1.9=h516909a_0
glib=2.58.3=py37h6f030ca_1002
gmp=6.1.2=hf484d3e_1000
gnutls=3.6.5=hd3a4fd2_1002
graphite2=1.3.13=hf484d3e_1000
gst-plugins-base=1.14.5=h0935bb2_0
gstreamer=1.14.5=h36ae1b5_0
harfbuzz=2.4.0=h9f30f68_3
hdf5=1.10.5=nompi_h3c11f04_1104
icu=64.2=he1b5a44_1
intel-openmp=2019.4=243
ipykernel=5.1.3=py37h39e3cac_0
ipython=7.10.1=py37h39e3cac_0
ipython_genutils=0.2.0=py37_0
jasper=1.900.1=h07fcdf6_1006
jedi=0.15.1=py37_0
jpeg=9c=h14c3975_1001
jupyter_client=5.3.4=py37_0
jupyter_core=4.6.1=py37_0
kiwisolver=1.1.0=pypi_0
lame=3.100=h14c3975_1001
libblas=3.8.0=14_openblas
libcblas=3.8.0=14_openblas
libclang=9.0.0=default_hde54327_4
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libiconv=1.15=h516909a_1005
liblapack=3.8.0=14_openblas
liblapacke=3.8.0=14_openblas
libllvm9=9.0.0=hc9558a2_3
libopenblas=0.3.7=h5ec1e0e_5
libopencv=4.1.2=py37_2
libpng=1.6.37=hbc83047_0
libprotobuf=3.11.1=h8b12597_0
libsodium=1.0.16=h1bed415_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_0
libuuid=2.32.1=h14c3975_1000
libwebp=1.0.2=h576950b_1
libxcb=1.13=h14c3975_1002
libxkbcommon=0.9.1=hebb1f50_0
libxml2=2.9.10=hee79883_0
matplotlib=3.1.2=pypi_0
mkl=2019.4=243
mkl-service=2.3.0=py37he904b0f_0
mkl_fft=1.0.15=py37h516909a_1
mkl_random=1.1.0=py37hb3f55d8_0
munch=2.5.0=pypi_0
ncurses=6.1=he6710b0_1
nettle=3.4.1=h1bed415_1002
networkx=2.4=py_0
ninja=1.9.0=py37hfd86e86_0
nspr=4.24=he1b5a44_0
nss=3.47=he751ad9_0
numpy=1.17.3=py37h95a1406_0
numpy-base=1.17.4=py37h2f8d375_0
olefile=0.46=py_0
opencv=4.1.2=py37_2
openh264=1.8.0=hdbcaa40_1000
openssl=1.1.1=h7b6447c_0
parso=0.5.1=py_0
pcre=8.43=he1b5a44_0
pexpect=4.7.0=py37_0
pickleshare=0.7.5=py37_0
pillow=6.2.1=py37h34e0f95_0
pip=19.3.1=py37_0
pixman=0.38.0=h516909a_1003
pretrainedmodels=0.7.4=pypi_0
prompt_toolkit=3.0.2=py_0
protobuf=3.11.1=py37he1b5a44_0
pthread-stubs=0.4=h14c3975_1001
ptyprocess=0.6.0=py37_0
py-opencv=4.1.2=py37h5ca1d4c_2
pycparser=2.19=py_0
pygments=2.5.2=py_0
pyparsing=2.4.5=pypi_0
python=3.7.5=h0371630_0
python-dateutil=2.8.1=py_0
pytorch=1.0.1=py3.7_cuda10.0.130_cudnn7.4.2_2
pyzmq=18.1.0=py37he6710b0_0
qt=5.12.5=hd8c4c69_1
readline=7.0=h7b6447c_5
scipy=1.3.2=py37he2b7bc3_0
setuptools=42.0.2=py37_0
six=1.13.0=py37_0
sqlite=3.30.1=h7b6447c_0
tensorboardx=1.9=py_0
tk=8.6.8=hbc83047_0
torchvision=0.2.2=py_3
tornado=6.0.3=py37h7b6447c_0
tqdm=4.40.2=pypi_0
traitlets=4.3.3=py37_0
wcwidth=0.1.7=py37_0
wheel=0.33.6=py37_0
x264=1!152.20180806=h14c3975_0
xorg-kbproto=1.0.7=h14c3975_1002
xorg-libice=1.0.10=h516909a_0
xorg-libsm=1.2.3=h84519dc_1000
xorg-libx11=1.6.9=h516909a_0
xorg-libxau=1.0.9=h14c3975_0
xorg-libxdmcp=1.1.3=h516909a_0
xorg-libxext=1.3.4=h516909a_0
xorg-libxrender=0.9.10=h516909a_1002
xorg-renderproto=0.11.1=h14c3975_1002
xorg-xextproto=7.3.0=h14c3975_1002
xorg-xproto=7.0.31=h14c3975_1007
xz=5.2.4=h14c3975_4
zeromq=4.3.1=he6710b0_3
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0
